GUACAMOLE-175: Move guacd_handle_user() to libguacd.
