+++ src/libguacd/libguacd/log.h
@@ -32,5 +32,11 @@
 void guacd_client_log_guac_error(guac_client* client,
         guac_client_log_level level, const char* message);
 
+/**
+ * Logs a reasonable explanatory message regarding handshake failure based on
+ * the current value of guac_error.
+ */
+void guacd_client_log_handshake_failure(guac_client* client);
+
 #endif
